refactor: move Executor and ExecutorConfig to evm module
