+++ config/importer-offline.env.local
@@ -1,4 +1,4 @@
-RUST_LOG=info,stratus::eth::block_miner=warn,stratus::eth::executor=warn,stratus::eth::evm::revm=warn,importer_offline::rx=off
+RUST_LOG=info,stratus::eth::block_miner=warn,stratus::eth::evm::executor=warn,stratus::eth::evm::revm=warn,importer_offline::rx=off
 
 CHAIN_ID=2008
 EVMS=8
