+++ src/bin/importer_offline.rs
@@ -17,6 +17,7 @@ use futures::try_join;
 use futures::StreamExt;
 use itertools::Itertools;
 use stratus::config::ImporterOfflineConfig;
+use stratus::eth::evm::Executor;
 use stratus::eth::primitives::Block;
 use stratus::eth::primitives::BlockNumber;
 use stratus::eth::primitives::ExternalBlock;
@@ -25,7 +26,6 @@ use stratus::eth::primitives::ExternalReceipts;
 use stratus::eth::storage::ExternalRpcStorage;
 use stratus::eth::storage::InMemoryPermanentStorage;
 use stratus::eth::BlockMiner;
-use stratus::eth::Executor;
 use stratus::ext::spawn_named;
 use stratus::ext::spawn_thread;
 use stratus::ext::to_json_string_pretty;
