+++ src/bin/importer_online.rs
@@ -9,6 +9,7 @@ use futures::StreamExt;
 use serde::Deserialize;
 use stratus::channel_read;
 use stratus::config::ImporterOnlineConfig;
+use stratus::eth::evm::Executor;
 use stratus::eth::primitives::BlockNumber;
 use stratus::eth::primitives::ExternalBlock;
 use stratus::eth::primitives::ExternalReceipt;
@@ -16,7 +17,6 @@ use stratus::eth::primitives::ExternalReceipts;
 use stratus::eth::primitives::Hash;
 use stratus::eth::storage::StratusStorage;
 use stratus::eth::BlockMiner;
-use stratus::eth::Executor;
 use stratus::ext::spawn_named;
 use stratus::ext::traced_sleep;
 use stratus::ext::DisplayExt;
