+++ src/config.rs
@@ -1,6 +1,5 @@
 //! Application configuration.
 
-use std::cmp::max;
 use std::net::SocketAddr;
 use std::str::FromStr;
 use std::sync::atomic::AtomicUsize;
@@ -15,6 +14,7 @@ use strum::VariantNames;
 use tokio::runtime::Builder;
 use tokio::runtime::Runtime;
 
+use crate::eth::evm::ExecutorConfig;
 #[cfg(feature = "dev")]
 use crate::eth::primitives::test_accounts;
 use crate::eth::primitives::Address;
@@ -35,8 +35,6 @@ use crate::eth::storage::StratusStorage;
 use crate::eth::storage::TemporaryStorage;
 use crate::eth::BlockMiner;
 use crate::eth::BlockMinerMode;
-use crate::eth::Executor;
-use crate::eth::ExecutorStrategy;
 use crate::eth::TransactionRelayer;
 use crate::ext::parse_duration;
 use crate::infra::build_info;
@@ -208,39 +206,6 @@ impl StratusStorageConfig {
         let storage = StratusStorage::new(temp_storage, perm_storage);
 
         Ok(Arc::new(storage))
-    }
-}
-
-// -----------------------------------------------------------------------------
-// Config: Executor
-// -----------------------------------------------------------------------------
-
-#[derive(Parser, DebugAsJson, Clone, serde::Serialize)]
-pub struct ExecutorConfig {
-    /// Chain ID of the network.
-    #[arg(long = "chain-id", env = "CHAIN_ID")]
-    pub chain_id: u64,
-
-    /// Number of EVM instances to run.
-    #[arg(long = "evms", env = "EVMS")]
-    pub num_evms: usize,
-
-    /// EVM execution strategy.
-    #[arg(long = "strategy", env = "STRATEGY", default_value = "serial")]
-    pub strategy: ExecutorStrategy,
-}
-
-impl ExecutorConfig {
-    /// Initializes Executor.
-    ///
-    /// Note: Should be called only after async runtime is initialized.
-    pub fn init(&self, storage: Arc<StratusStorage>, miner: Arc<BlockMiner>) -> Arc<Executor> {
-        let mut config = self.clone();
-        config.num_evms = max(config.num_evms, 1);
-        tracing::info!(?config, "creating executor");
-
-        let executor = Executor::new(storage, miner, config);
-        Arc::new(executor)
     }
 }
