+++ src/eth/evm/mod.rs
@@ -1,9 +1,15 @@
 #[allow(clippy::module_inception)]
 mod evm;
 mod evm_error;
-pub mod revm;
+mod executor;
+mod executor_config;
+mod revm;
 
 pub use evm::Evm;
 pub use evm::EvmExecutionResult;
 pub use evm::EvmInput;
 pub use evm_error::EvmError;
+pub use executor::Executor;
+pub use executor::ExecutorStrategy;
+pub use executor_config::ExecutorConfig;
+pub use revm::Revm;
