+++ src/eth/rpc/rpc_server.rs
@@ -27,6 +27,7 @@ use tracing::info_span;
 use tracing::Instrument;
 use tracing::Span;
 
+use crate::eth::evm::Executor;
 use crate::eth::primitives::Address;
 use crate::eth::primitives::BlockFilter;
 #[cfg(feature = "dev")]
@@ -52,7 +53,6 @@ use crate::eth::rpc::RpcSubscriptions;
 use crate::eth::storage::StratusStorage;
 use crate::eth::BlockMiner;
 use crate::eth::Consensus;
-use crate::eth::Executor;
 use crate::ext::not;
 use crate::ext::to_json_string;
 use crate::ext::to_json_value;
